Project location edit test case, moved add location test case
