+++ functional_tests/selenium_tests/tests/__init__.py
@@ -51,13 +51,16 @@ from functional_tests.selenium_tests.tests.projects.project import (
 )
 from functional_tests.selenium_tests.tests.projects.project_draw_on_map import (
     AddProjectWithExtent,
-    AddLocation,
 )
 from functional_tests.selenium_tests.tests.projects.search_project import (
     ProjectSearch
 )
 from functional_tests.selenium_tests.tests.projects.project_resources import (
     AddResource
+)
+from functional_tests.selenium_tests.tests.projects.project_location import (
+    AddLocation,
+    EditLocation
 )
 from functional_tests.selenium_tests.tests.projects.location_resources import (
     AddLocationResource
